clk: keystone: sci-clk: set exact clock rate first before using tolerance
When requesting to set clock rate for a device, it is not always set
to exact value. So first try to set the rate to the desired value and
if that returns an error, try with the given tolerance.
'determine-rate' modifies the request for set-rate hook. The requested
value is changed to the closest tolerable value. So do the same there.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Tested-by: Aradhya Bhatia <a-bhatia1@ti.com>